Move global functions into namespace
