+++ plugins/core/IO/qLASIO/include/CMakeLists.txt
@@ -6,6 +6,8 @@ target_sources(${PROJECT_NAME}
         ${CMAKE_CURRENT_LIST_DIR}/LasDetails.h
         ${CMAKE_CURRENT_LIST_DIR}/LasOpenDialog.h
         ${CMAKE_CURRENT_LIST_DIR}/LasSaveDialog.h
+        ${CMAKE_CURRENT_LIST_DIR}/LasScalarField.h
+        ${CMAKE_CURRENT_LIST_DIR}/LasExtraScalarField.h
         ${CMAKE_CURRENT_LIST_DIR}/LasScalarFieldLoader.h
         ${CMAKE_CURRENT_LIST_DIR}/LasScalarFieldSaver.h
         ${CMAKE_CURRENT_LIST_DIR}/LasWaveformLoader.h
