+++ plugins/core/IO/qLASIO/include/LasOpenDialog.h
@@ -18,6 +18,8 @@
 //##########################################################################
 
 #include "LasDetails.h"
+#include "LasScalarField.h"
+#include "LasExtraScalarField.h"
 
 // GUI generated by Qt Designer
 #include <ui_lasopendialog.h>
