+++ plugins/core/IO/qLASIO/include/LasSavedInfo.h
@@ -18,6 +18,7 @@
 //##########################################################################
 
 #include "LasDetails.h"
+#include "LasExtraScalarField.h"
 
 #include <laszip/laszip_api.h>
