+++ plugins/core/IO/qLASIO/src/CMakeLists.txt
@@ -5,6 +5,8 @@ target_sources(${PROJECT_NAME}
         ${CMAKE_CURRENT_LIST_DIR}/LasIOFilter.cpp
         ${CMAKE_CURRENT_LIST_DIR}/LasOpenDialog.cpp
         ${CMAKE_CURRENT_LIST_DIR}/LasSaveDialog.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/LasScalarField.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/LasExtraScalarField.cpp
         ${CMAKE_CURRENT_LIST_DIR}/LasDetails.cpp
         ${CMAKE_CURRENT_LIST_DIR}/LasScalarFieldLoader.cpp
         ${CMAKE_CURRENT_LIST_DIR}/LasScalarFieldSaver.cpp
