+++ plugins/core/IO/qLASIO/src/LasWaveformSaver.cpp
@@ -34,7 +34,7 @@ void LasWaveformSaver::handlePoint(size_t index, laszip_point &point)
         QDataStream stream(&m_array, QIODevice::WriteOnly);
         stream.setByteOrder(QDataStream::ByteOrder::LittleEndian);
         stream << w.descriptorID();
-        stream << static_cast<quint64>(w.dataOffset() + EvlrHeader::SIZE);
+        stream << static_cast<quint64>(w.dataOffset() + LasDetails::EvlrHeader::SIZE);
         stream << w.byteCount();
         stream << w.echoTime_ps();
         stream << w.beamDir().x << w.beamDir().y << w.beamDir().z;
